requirements.txt
alembic==1.4.0
aniso8601==8.0.0
certifi==2019.11.28
Click==7.0
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.21.0
packaging==20.1
pipenv==2018.11.26
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
toml==0.10.0
tox==3.14.3
virtualenv==16.7.9
virtualenv-clone==0.5.3
Werkzeug==1.0.0
zipp==2.2.0
